+++ src/assassin/feeds/discount_option_data.rs
@@ -65,29 +65,29 @@ impl DataFeed for DiscountOptionData {
 			Utc.ymd(date_parts[0] as i32, date_parts[1], date_parts[2]).and_hms(0, 0, 0)
 		};
 
-		let ask: f64 = v[2].parse().unwrap();
-		let bid: f64 = v[4].parse().unwrap();
+		let ask: f32 = v[2].parse().unwrap();
+		let bid: f32 = v[4].parse().unwrap();
 
 		// TODO: this might be legit in certain market situations...
 		if bid > ask {
 			panic!("bid {} is > ask {}", bid, ask);
 		}
 
-		let last_price: f64 = v[6].parse().unwrap();
+		let last_price: f32 = v[6].parse().unwrap();
 		let call: String = v[7].parse().unwrap();
 
 		if call != "call" && call != "put" {
 			panic!("expected 'call' or 'put', got {}", call);
 		}
 
-		let strike_price: f64 = v[8].parse().unwrap();
+		let strike_price: f32 = v[8].parse().unwrap();
 		let volume: i32 = v[9].parse().unwrap();
-		let implied_volatility: f64 = v[10].parse().unwrap();
-		let delta: f64 = v[11].parse().unwrap();
-		let gamma: f64 = v[12].parse().unwrap();
-		let vega: f64 = v[13].parse().unwrap();
+		let implied_volatility: f32 = v[10].parse().unwrap();
+		let delta: f32 = v[11].parse().unwrap();
+		let gamma: f32 = v[12].parse().unwrap();
+		let vega: f32 = v[13].parse().unwrap();
 		let open_interest: i32 = v[15].parse().unwrap();
-		let underlying_price: f64 = v[16].parse().unwrap();
+		let underlying_price: f32 = v[16].parse().unwrap();
 
 		let date = {
 			let date_parts: Vec<u32> = v[17].split('-').map(|p| p.parse().unwrap()).collect();
